# Conflicts:
#	CHANGELOG.md
#	Verovio.podspec
#	bindings/java/pom.xml
#	bindings/python/setup.py

+++ include/vrv/functorparams.h
@@ -1199,12 +1199,12 @@ public:
         m_currentTempo = 120;
         m_functor = functor;
     }
-    std::map<int, double> realTimeToScoreTime;
-    std::map<int, std::vector<std::string> > realTimeToOnElements;
-    std::map<int, std::vector<std::string> > realTimeToOffElements;
-    std::map<int, int> realTimeToTempo;
+    std::map<double, double> realTimeToScoreTime;
+    std::map<double, std::vector<std::string> > realTimeToOnElements;
+    std::map<double, std::vector<std::string> > realTimeToOffElements;
+    std::map<double, int> realTimeToTempo;
     double m_scoreTimeOffset;
-    int m_realTimeOffsetMilliseconds;
+    double m_realTimeOffsetMilliseconds;
     int m_currentTempo;
     Functor *m_functor;
 };
